Fix exception thrown after updating framework from 3.1.1 to 3.1.2
SearchVariantVersioned makes modifications to the $manipulation array
constructed during a DataObject::write call. One of these modifications
involves stripping "_Live" from table names so that they can be used as
class names. Before modifications specific to the Versioned extension
SearchVariantVersioned::appliesTo() strictly checks if the object being
are made, saved has the extension versioned. Since this check **explicitly
excludes subclasseses of Versioned**, the check fails for classes with the
VersionedDataObject extension, and "_Live" isn't stripped out of the table
name variables. The result of this is an exception:

    InvalidArgumentException: Slice_Live is not a subclass of DataObject