Remove incorrect parameter passed to FormAction constructor
The third parameter of the FormAction constructor is supposed to be an
instance of Form. The string being passed in this case was effectively
being ignored as it was set as the value in the parent class constructor.
The commit 1db08ba in framework changed the body of the FormAction
constructor to pass the third parameter into FormAction::setForm, which
causes an error if a string value is used like it was here.

https://github.com/silverstripe/silverstripe-framework/commit/1db08bac88f9330dc4e6dda1ae08628f245a5212

Resolves #6