Fix incorrect write behaviour when stage viewing mode is set to Live
When the stage reading mode was set to Live and the "Save draft" action was
used, the form submission would be written directly to the live table. As
the "save and publish" action copies a stage record to the live table, it
was possible for the following problem to occur:

1. Select the "Stage" viewing mode
2. Save and publish record X: works correctly
3. Select the "Live" viewing mode
4. Change record X and save as draft: writes to live, stage not changed
5. Save and publish record X: copies old draft (from 2.) over newer live version
6. (anguish because all changes since the record was saved from the "stage"
   reading mode are lost forever)

In SilverStripe 3.1.2, the way that `Versioned`'s "stage" vs. "live" is
handled changed, which causes `DataObject::write()` to write to the current
"reading stage" rather than always to "Stage" as it used to (presumably to
make the decorator more generic, but it appears a bit confused).

http://doc.silverstripe.org/en/changelogs/3.1.2/